requirements.txt
google-cloud-vision==0.31.0
python-dotenv==0.8.2
SimpleCV==1.3.0
